backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-jose==3.5.0
passlib[bcrypt]==1.7.4
